requirements.txt
# Telegram Bot
aiogram==3.13.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
aiosqlite==0.20.0
greenlet==3.1.1

# API
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Google Sheets Parser
gspread==6.1.4
google-auth==2.36.0

# Utils
python-dotenv==1.0.1

